requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.8
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
emails==0.6
boto3==1.34.0
stripe==7.8.0
requests==2.31.0
pillow==10.1.0
aiofiles==23.2.1
httpx==0.25.2
pydantic[email]==2.5.0
alembic==1.13.0
python-dateutil==2.8.2